refactor: remove the legacy query cascade and conductor-DB code
The block system and conductor state resolve through the holochain_data
ConductorStore, and the cascade reads the DhtStore directly, so the legacy
holochain_state query transaction-cascade layer and the conductor-database
helpers are dead.

Remove the Query/Stores/CascadeTxnWrapper/DbScratch machinery and its
per-query submodules (chain_head, entry_details, link_count, link_details,
live_entry, live_record, record_details), the conductor-DB block helpers
(block/unblock/insert_block) and query_conductor_state, and the
authored/DHT/conductor/wasm test-database helpers. The live pieces are
kept: StateQueryError, GetLinksFilter, from_blob/to_blob, and the Store
trait used by the DhtStore scratch reads.

With this, no code outside holochain_sqlite references any DbKind type.

Part of #5722.
