+++ crates/holochain_state/src/lib.rs
@@ -2,8 +2,8 @@
 //! with the `holochain_sqlite` crate.
 //!
 //! ## Reads
-//! The main abstraction for creating data read queries is the [`Query`](crate::query::Query) trait.
-//! This can be implemented to make constructing complex queries easier.
+//! The [`DhtStore`] and [`DhtStoreRead`] types are the main abstractions for
+//! reading data, combining database access with the in-memory scratch space.
 //!
 //! The [`source_chain`] module provides the [`SourceChain`](crate::source_chain::SourceChain) type,
 //! which is the abstraction for working with chains of actions.
@@ -20,13 +20,9 @@
 //!
 //! The SourceChain type uses the Scratch for in-memory operations which
 //! can be flushed to the database.
-//!
-//! The Query trait allows combining arbitrary database SQL queries with
-//! the scratch space so reads can union across the database and in-memory data.
 
 pub use dht_store::{DhtStore, DhtStoreRead};
 
-pub mod block;
 pub mod chain_lock;
 pub mod conductor;
 pub mod dht_store;
